+++ src/main/java/frc/robot/constants/ShooterConstants.java
@@ -1,6 +1,6 @@
 package frc.robot.constants;
 
-public class OuttakeConstants {
+public class ShooterConstants {
 
     // motor Ids (99 is placeholder for now)
     public static final int BOTTOM_MOTOR_ID = 99;
