+++ src/main/java/frc/robot/subsystems/gpm/Intake.java
@@ -1,0 +1,5 @@
+package frc.robot.subsystems.gpm;
+
+public class Intake {
+    
+}
