--- src/main/java/frc/robot/subsystems/gpm_subsystem/Intake.java
@@ -1,5 +1,0 @@
-package frc.robot.subsystems.gpm_subsystem;
-
-public class Intake {
-    
-}
